+++ src/app.tsx
@@ -126,6 +126,7 @@ const RoomActions: FC<{identity: Identity, roomId: string}> = ({identity, roomId
             <li><a href={`#/${encodeURIComponent(identity.name)}/${encodeURIComponent(roomId)}/invite`}>Bulk invite</a></li>
             <li><a href={`#/${encodeURIComponent(identity.name)}/${encodeURIComponent(roomId)}/kick`}>Bulk kick</a></li>
             <li><a href={`#/${encodeURIComponent(identity.name)}/${encodeURIComponent(roomId)}/mass-joiner`}>Mass joiner (AppService API)</a></li>
+            <li><a href={`#/${encodeURIComponent(identity.name)}/${encodeURIComponent(roomId)}/live-location-sharing`}>Live Location Sharing</a></li>
             <li><a href={`#/${encodeURIComponent(identity.name)}/${encodeURIComponent(roomId)}/yaml`}>JSON export</a></li>
         </ul></nav>
     </>;